@alvacouch @pkdash @dtarb @mjstealey I have made the following changes and did the timing using the zip file in this resource https://www.hydroshare.org/resource/37594b5bba544e14905dc0f8257c874f/ as the test case. This zip file include many folders and a large set of files. Note that ibun unzip call only took 4 seconds for this test case, which can be ignored compared to the link_irods_folder_to_django() call which takes 102 seconds as is now.

1. removed the deprecated istorage.size() call. This reduced the link call from 102 seconds to 82 seconds.
2. removed duplicate ```istorage = resource.get_irods_storage()``` call by adding the if condition to only call it when pass in istorage is None. Since we already pass in an istorage object, this call is not needed. This reduced the link call from 82 seconds to 81 seconds consistently, a small improvement, although not critical, but nice to have.
3. removed istorage.exist() verification calls in ResourceFile.create() call for this specific case where we already know files exist in iRODS as the file names come from icommand list dir ils call, so no need to call istorage.exist() to verify file exist in iRODS. This further reduced the time from 81 seconds to 62 seconds.

So this PR if merged will reduce the unzip time from the original 106 seconds (4 second ibun call plus 102 seconds link call) to 66 seconds (4 second ibun call plus 62 seconds link call) for this test case. Please take a look at this PR @alvacouch and @pkdash at your earliest convenience to see if you can +1 this so that we can merge it for Tanu and her team to test tomorrow to see if the performance is acceptable for their demo for Earthcube workshop that will happen this Wed. If you approve this PR. I'll create another PR against the hotfix branch. @pkdash @mjstealey Please advise which branch I should create the PR against for them to use. I assume they will do demo against master branch, right?