@dtarb 2017-06-06
Thanks.  This will help a bit.  108 to 66 seconds is a worthwhile improvement, though still rather long.

I have to ask the question as to why we need to store a record of each file in Django.  For resources that have lots of files, as an increasing number do, could we not just rely on the iRODS listing for these.  The UI already does this.  Could you all please weigh in on the feasibility and complexity of this, and if this seems a worthwhile simplification, could @alvacouch please drive a methodical switch to this (not as a hotfix, but in an orderly way for a future release at a good time).