@hyi 2017-06-06
@dtarb @alvacouch @mjstealey @pkdash @aphelionz I took a look at the master branch and realized we still have the old file system implementation on master, so only part of the first performance improvement I made, i.e., remove the istorage.size() call when it is not needed, applied to the master branch. I don't anticipate enough performance improvement by doing that based on the timing tests I did for the new file system implementation, so I don't think applying a hotfix on master for this is worthwhile at this point. Regardless, this PR is nice to have as a hotfix for release 1.10. I'll make a PR against release candidate 1.10 after this PR is approved to get merged to develop. @pkdash told me he is out of the office until the coming Monday, so he is not able to review this PR. So I am hoping @alvacouch or @aphelionz can review this when you have a chance so it can make its way into our next release 1.10. Thanks.