@alvacouch 2017-06-07
@hyi @dtarb There are many reasons we list files in Django. 

1. It makes a sanity check of iRODS.  Without this, we would not have known there were problems. Ignorance is not bliss. 
2. It allows easy generation of metadata files without lots of interaction with irods. 
3. It allows us to remember what was there if people change things. This will allow us to change metadata appropriately and efficiently if -- e.g. -- there are large changes in the kinds of files, e.g., via unzip.