@aphelionz 2017-06-07
Can we perhaps pool some of the interactions so we're not doing one per file? Can we do a number of inserts at once to save time?