@alvacouch 2017-06-07
@hyi @dtarb This change does not prevent unzip operations that create duplicate files in the ResourceFile hierarchy. While this is unlikely, it is possible when the unzipped file goes into an existing directory.