@alvacouch 2017-06-07
@hyi @dtarb changing over to iRODS as a ground truth for files is inconsistent with the file metadata model for composite resource, and as well, makes it difficult to manage file upload via native Django mechanisms.