@hyi 2017-06-07
@dtarb @alvacouch I agree with @alvacouch on his comments especially regarding making it difficult to manage file upload via native Django mechanisms if we were to remove ResourceFile object for keeping track of files in iRODS.