@hyi 2017-06-07
@aphelionz I thought about the same thing regarding your comment to pooling some of the interactions so we're not doing one per file but do a number of inserts at once to save time, but I cannot come up with a simple solution to achieve this. I'll merge this PR now to get us at least some performance gain. If you have specific ideas to pursue further performance improvement, please go ahead with it.